+++ tests/async_combinators_test.clj
@@ -181,4 +181,4 @@
       (drain!)
       (is (= [10 20 30 40 50 60 70 80] @result)))))
 
-(run-tests)
+(run-tests-and-exit)